The Anatomy of CrowdStrike Threat Hunting
I wrote this capstone report for a culmination of learning within Federated Hermes' Information Security Group provided for a presentation be completed on a specified topic. 

I spent days watching videos, using the tools, and understanding the concepts behind the scenes to provide the company with an inclusive of CrowdStrike's Threat Hunting capabilities.